+++ models/user.js
@@ -1,10 +1,6 @@
 const mongoose = require("../db/connection");
 const { Schema } = mongoose;
-
-const validateEmail = (email) => {
-  const regex = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
-  return regex.test(email);
-};
+const validateEmail = require("../utils/validate-email");
 
 const UserModel = mongoose.model(
   "User",
