+++ utils/validate-email.js
@@ -1,0 +1,6 @@
+const validateEmail = (email) => {
+  const regex = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
+  return regex.test(email);
+};
+
+module.exports = validateEmail
